GEO-41 #time 1h Resolving issues for the integration of OpenLayers base maps from Islandora GeoTIFF's

+++ islandora_gis_geoserver/libraries/Shapefile.php
@@ -21,35 +21,72 @@ class IslandoraGeoImage extends IslandoraLargeImage {
 
   public $geoserver_session;
   public $shapefiles;
+  public $coverage_name;
 
   // hasDerivation
   // GeoServerCoverageStore
   function __construct($islandora_session,
-		       $geoserver_session,
-		       $pid = NULL, $object = NULL) {
+		       $geoserver_session = NULL,
+		       $pid = NULL,
+		       $workspace = 'default',
+		       $object = NULL) {
 
     parent::__construct($islandora_session, $pid, $object);
 
-    $coveragestore_name = preg_replace('/\:/', '_', $this->object->id);
+    //$this->coverage_name = preg_replace('/\:/', '_', $this->object->id);
+    $this->coverage_name = preg_replace('/\:/', '_', $pid);
 
-    $this->client = new IslandoraGeoServerClient($geoserver_session);
-    $workspace = $this->client->workspace('default');
+    if(!is_a($this->object, 'FedoraObject')) {
 
-    //$file_path = '/var/www/drupal/sites/all/modules/islandora_dss_solution_pack_gis/islandora_gis_geoserver/eapl-sanborn-easton-1919_010_modified.tif';
-    $file_path = '/tmp/IslandoraGeoImage_' . $coveragestore_name . '.tiff';
+      throw new Exception("Failed to pass a tuque FedoraObject to the IslandoraGeoImage constructor");
+    }
 
-    $ds_obj = $this->object['OBJ'];
-    $ds_obj->getContent($file_path);
+    if(!is_null($geoserver_session)) {
 
-    $workspace->createCoverageStore($coveragestore_name, $file_path);
-    unlink($file_path);
+      $this->client = new IslandoraGeoServerClient($geoserver_session);
+      $workspace = $this->client->workspace($workspace);
+
+      //$file_path = '/var/www/drupal/sites/all/modules/islandora_dss_solution_pack_gis/islandora_gis_geoserver/eapl-sanborn-easton-1919_010_modified.tif';
+      $file_path = '/tmp/IslandoraGeoImage_' . $this->coverage_name . '.tiff';
+
+      $ds_obj = $this->datastream('OBJ');
+      $ds_obj->getContent($file_path);
+
+      $workspace->createCoverageStore($this->coverage_name, $file_path);
+      unlink($file_path);
+    }
   }
 
   function add_shapefile($shapefile) {
 
-    $shapefile_uri = 'info:fedora/' . $shapefile->id;
+    $shapefile_uri = $shapefile->id;
 
     $this->object->relationships->add(FEDORA_RELS_EXT_URI, 'hasDependent', $shapefile_uri);
+  }
+
+  function to_layer() {
+
+    $layer = new stdClass();
+    $layer->name = $this->coverage_name;
+    $layer->title = $this->object->label;
+
+    $layer->data = array('openlayers' => array('gwc_data' => array('isBaseLayer' => TRUE,
+								   'projection' => array('EPSG:900913', 'EPSG:3857'))));
+
+    return $layer;
+    /*
+    $openlayers = $gs_layer->data['openlayers'];
+
+    if ($openlayers['gwc']) {
+      $data = isset($openlayers['gwc_data']) ? $openlayers['gwc_data'] : array();
+     */
+
+    /*
+        'isBaseLayer' => isset($data['isBaseLayer']) ? $data['isBaseLayer'] : FALSE,
+        'projection' => isset($data['projection']) ? $data['projection'] : array('EPSG:900913', 'EPSG:3857'),
+        'params' => array(
+          'isBaseLayer' => isset($data['isBaseLayer']) ? $data['isBaseLayer'] : FALSE,
+     */
   }
 }
 
@@ -60,15 +97,16 @@ class IslandoraGeoImage extends IslandoraLargeImage {
 
 class IslandoraShapefile extends IslandoraObject {
 
+  public $geoserver_session;
   public $base_maps;
 
   function __construct($session,
-		       $geoserver_session,
+		       $geoserver_session = NULL,
 		       $pid = NULL, $object = NULL) {
 
     parent::__construct($session, $pid, $object);
 
-    $this->geoserver_session;
+    $this->geoserver_session = $geoserver_session;
 
     $this->base_maps = array();
     $this->get_base_maps();
@@ -84,13 +122,14 @@ class IslandoraShapefile extends IslandoraObject {
               <fedora-model:hasModel> $content ;
               <fedora-model:state> <fedora-model:Active> .';
     $query .= '} ORDER BY $title';
+
     $query_array = array(
 			 'query' => $query,
 			 'type' => 'sparql',
 			 //'pid' => $obj_pid,
 			 // Seems as though this is ignored completely.
-			 'page_size' => $page_size,
-			 'page_number' => $page_number,
+			 //'page_size' => $page_size,
+			 //'page_number' => $page_number,
 			 );
 
     foreach($this->session->connection->repository->ri->query($query_array['query'], $query_array['type']) as $result) {
@@ -98,5 +137,4 @@ class IslandoraShapefile extends IslandoraObject {
       $this->base_maps[] = new IslandoraGeoImage($this->session, $this->geoserver_session, $result['object']['value']);
     }
   }
-
-  }
+}
